+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JWVWDD1YG087PMN6T7SW9CEM"
+        "01JWWTBJG8WYBTFFS654S71A5V"
     )
     r"""The id of the resource"""