+++ packages/orq-rc/src/orq_ai_sdk/models/retrievecontactop.py
@@ -96,7 +96,7 @@ class RetrieveContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-06-03T21:50:51.469Z")
+    updated: Optional[datetime] = parse_datetime("2025-06-04T06:34:07.791Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")