+++ packages/orq-rc/RELEASES.md
@@ -1669,3 +1669,13 @@ Based on:
 - [python v3.7.0-rc.9] packages/orq-rc
 ### Releases
 - [PyPI v3.7.0-rc.9] https://pypi.org/project/orq-ai-sdk/3.7.0-rc.9 - packages/orq-rc
+
+## 2025-06-04 06:35:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.555.2 (2.620.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.8.0-rc.0] packages/orq-rc
+### Releases
+- [PyPI v3.8.0-rc.0] https://pypi.org/project/orq-ai-sdk/3.8.0-rc.0 - packages/orq-rc